feat: move show_map to visualization and allow Ylm as arg
