+++ src/jaxoplanet/experimental/starry/__init__.py
@@ -3,3 +3,5 @@ from jaxoplanet.experimental.starry.solution import (
     solution_vector as solution_vector,
 )
 from jaxoplanet.experimental.starry.ylm import Ylm as Ylm
+
+from jaxoplanet.experimental.starry.vizualisation import show_map as show_map
